requirements.txt
absl-py==1.1.0
altair==4.2.0
altgraph==0.17
appdirs==1.4.3
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.4.1
asttokens==2.0.5
astunparse==1.6.3
attrs==21.2.0
audioread==2.1.9
backcall==0.2.0
backports.entry-points-selectable==1.1.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.10.0
bleach==5.0.0
blinker==1.4
Brlapi==0.7.0
cachetools==5.2.0
certifi==2019.11.28
cffi==1.15.0
chardet==3.0.4
chrome-gnome-shell==0.0.0
Click==7.0
colorama==0.4.3
command-not-found==0.3
commonmark==0.9.1
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
debugpy==1.6.0
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.2
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==0.5.0
Django==3.2.9
django-heroku==0.0.0
docutils==0.18.1
duplicity==0.8.12.0
entrypoints==0.3
evdev==1.4.0
executing==0.8.3
fasteners==0.14.1
fastjsonschema==2.15.3
filelock==3.0.12
flatbuffers==1.12
future==0.18.2
gast==0.4.0
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gpg===1.13.1-unknown
grpcio==1.46.3
gunicorn==20.1.0
h5py==3.7.0
httplib2==0.14.0
idna==2.8
importlib-metadata==4.11.4
importlib-resources==5.7.1
ipykernel==6.13.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.6.0
jupyter-client==7.3.3
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
keras==2.9.0
Keras-Preprocessing==1.1.2
keyring==18.0.1
kiwisolver==1.3.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libclang==14.0.1
librosa==0.9.1
llvmlite==0.38.1
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib==3.4.2
matplotlib-inline==0.1.3
mediapipe==0.8.6
mistune==0.8.4
mock==3.0.5
monotonic==1.5
more-itertools==4.2.0
MouseInfo==0.1.3
nbclient==0.6.4
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
netifaces==0.10.4
notebook==6.4.12
numba==0.55.2
numpy==1.21.0
oauthlib==3.1.0
olefile==0.46
opencv-contrib-python==4.5.3.56
opencv-python==4.5.3.56
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
paramiko==2.6.0
parso==0.8.3
pbr==5.6.0
pexpect==4.6.0
pickle5==0.0.11
pickleshare==0.7.5
Pillow==7.0.0
pipenv==2022.7.4
platformdirs==2.1.0
pooch==1.6.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.17.3
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
PyAutoGUI==0.9.53
pycairo==1.16.2
pycparser==2.21
pycups==1.9.73
pydeck==0.7.1
pydub==0.25.1
pygame==2.1.2
PyGetWindow==0.0.9
Pygments==2.12.0
PyGObject==3.36.0
pyinstaller==4.4
pyinstaller-hooks-contrib==2021.2
PyJWT==1.7.1
pymacaroons==0.13.0
Pympler==1.0.1
PyMsgBox==1.0.9
PyNaCl==1.3.0
pynput==1.7.3
pyparsing==2.4.7
pyperclip==1.8.2
PyQt5==5.14.1
PyRect==0.1.4
pyRFC3339==1.1
pyrsistent==0.18.1
PyScreeze==0.1.27
PySocks==1.6.8
pytesseract==0.3.8
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-decouple==3.5
python-xlib==0.31
python3-xlib==0.15
PyTweening==1.0.3
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pyxdg==0.26
PyYAML==5.3.1
pyzmq==23.1.0
QtPy==1.9.0
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
resampy==0.2.2
rich==12.4.4
rsa==4.8
scikit-learn==1.1.1
scipy==1.8.1
SecretStorage==2.3.1
semver==2.13.0
Send2Trash==1.8.0
simplejson==3.16.0
sip==4.19.21
six==1.14.0
smmap==5.0.0
sounddevice==0.4.4
SoundFile==0.10.3.post1
soupsieve==2.3.1
sqlparse==0.4.2
stack-data==0.2.0
statistics==1.0.3.5
stevedore==3.3.0
streamlit==1.10.0
systemd-python==234
termcolor==1.1.0
terminado==0.15.0
threadpoolctl==3.1.0
tinycss2==1.1.1
toml==0.10.2
toolz==0.11.2
torbrowser-launcher==0.3.2
tornado==6.1
traitlets==5.2.2.post1
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
ubuntu-advantage-tools==27.9
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
validators==0.20.0
virtualenv==20.6.0
virtualenv-clone==0.5.5
wadllib==1.3.3
watchdog==2.1.8
wavio==0.0.4
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.1.2
whitenoise==5.3.0
widgetsnbextension==3.6.0
wrapt==1.14.1
xkit==0.0.0
zipp==3.8.0
